fix: 입력 값 파싱 기능 분리
- 입력 받아서 "," 기준으로 스플릿 하는 기능 Validator 클래스로 이관
- Parse 기능이 추가되어 Validator 클래스 이름 Parser로 변경
- parse 기능 유닛 테스트 추가
