--- src/main/java/com/nextstep/racingcar/Validator.java
@@ -1,9 +1,0 @@
-package com.nextstep.racingcar;
-
-public class Validator {
-    private static final int NAME_LIMIT = 5;
-
-    public static boolean isInvalidName(String name){
-        return NAME_LIMIT < name.length();
-    }
-}
